LPS-93662 Move OAuth2ProviderTestUtil to where it is used.
