--- modules/apps/oauth2-provider/oauth2-provider-test-util/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay OAuth2 Provider Test Utilities
-Bundle-SymbolicName: com.liferay.oauth2.provider.test.util
-Bundle-Version: 3.0.0
