--- modules/apps/oauth2-provider/oauth2-provider-test-util/build.gradle
@@ -1,14 +1,0 @@
-dependencies {
-	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.jboss.arquillian.container", name: "arquillian-container-test-spi", version: "1.1.10.Final"
-	compileOnly group: "org.jboss.arquillian.test", name: "arquillian-test-api", version: "1.2.0.Final"
-	compileOnly group: "org.jboss.shrinkwrap", name: "shrinkwrap-api", version: "1.2.2"
-	compileOnly group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
-	compileOnly project(":core:petra:petra-string")
-	compileOnly project(":test:shrinkwrap-osgi")
-}
-
-liferay {
-	deployDir = file("${liferayHome}/osgi/test")
-}
