+++ modules/apps/oauth2-provider/oauth2-provider-test/build.gradle
@@ -13,11 +13,15 @@ dependencies {
 	compileOnly project(":apps:static:osgi:osgi-util")
 	compileOnly project(":core:petra:petra-string")
 
+	testIntegrationCompile group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	testIntegrationCompile group: "com.liferay.arquillian", name: "com.liferay.arquillian.arquillian-container-liferay", version: "1.0.7"
 	testIntegrationCompile group: "com.liferay.arquillian", name: "com.liferay.arquillian.arquillian-deployment-generator-bnd", version: "1.0.4"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
 	testIntegrationCompile group: "org.apache.cxf", name: "cxf-rt-rs-client", version: "3.2.5"
+	testIntegrationCompile group: "org.jboss.arquillian.container", name: "arquillian-container-test-spi", version: "1.1.10.Final"
 	testIntegrationCompile group: "org.jboss.arquillian.junit", name: "arquillian-junit-container", version: "1.1.10.Final"
+	testIntegrationCompile group: "org.jboss.arquillian.test", name: "arquillian-test-api", version: "1.2.0.Final"
+	testIntegrationCompile group: "org.jboss.shrinkwrap", name: "shrinkwrap-api", version: "1.2.2"
 	testIntegrationCompile project(":apps:document-library:document-library-api")
-	testIntegrationCompile project(":apps:oauth2-provider:oauth2-provider-test-util")
+	testIntegrationCompile project(":test:shrinkwrap-osgi")
 }
